% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_data.R
\name{augment_data}
\alias{augment_data}
\title{Internal helper function for reformatting external and target data}
\usage{
augment_data(xtar, ytar, xext, yext, family, L = NULL, standardize = TRUE)
}
\arguments{
\item{xtar}{Predictor matrix from the target population}

\item{ytar}{Outcome vector from the target population}

\item{xext}{Predictor matrix from the external population}

\item{yext}{Outcome vector from the external population}

\item{family}{One of 'gaussian', 'binomial', or 'cox'}

\item{L}{Square matrix that scales the bias parameter gamma. If \code{NULL}, then
the \code{make_L()} function will be called to set \code{L} as the cholesky
decomposition of \code{crossprod(xtar)}}

\item{standardize}{If TRUE, then xext will be standardized but only in the
part of the augmented data matrix corresponding to gamma.}
}
\value{
A named list to be used by deglr
}
\description{
Internal helper function for reformatting external and target data
}
