% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_ploglik.R
\name{coxph_ploglik}
\alias{coxph_ploglik}
\title{Calculate the Cox partial log-likelihood}
\usage{
coxph_ploglik(y, x)
}
\arguments{
\item{y}{A vector of class Surv()}

\item{x}{A vector with length equal to that of y or matrix with number of rows
equal to that of y. Each element corresponds to a linear predictor.}
}
\value{
If x is a vector, then a single number corresponding to the partial
log-likelihood. If x is a matrix, then a vector corresponding to the partial
log-likelihoods for each column in x.
}
\description{
Calculate the Cox partial log-likelihood
}
