% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_deglr.R
\name{cv_deglr}
\alias{cv_deglr}
\title{Use cross-validation to select appropriate value of the tuning parameter in data-enriched generalized linear regression}
\usage{
cv_deglr(
  xtar,
  ytar,
  xext,
  yext,
  family,
  alpha = 0,
  nfolds = 5,
  ncvreps = 2,
  nlambda = 20,
  lambda.min.ratio = ifelse(nrow(xtar) < d/2, 0.001, 1e-06),
  lambda = NULL,
  L = NULL,
  standardize = TRUE,
  thresh = 1e-14,
  maxit = 1e+05
)
}
\arguments{
\item{xtar}{Predictor matrix from the target population}

\item{ytar}{Outcome vector from the target population}

\item{xext}{Predictor matrix from the external population}

\item{yext}{Outcome vector from the external population}

\item{family}{One of 'gaussian', 'binomial', or 'cox'}

\item{alpha}{See \code{glmnet::cv.glmnet()}}

\item{nfolds}{See \code{glmnet::cv.glmnet()}}

\item{ncvreps}{Number of partitions to create. Choosing > 1 improves
stability of selection but takes longer}

\item{nlambda}{See \code{glmnet::glmnet()} but note that in contrast to glmnet() the
sequence of lambdas will always be exactly this long}

\item{lambda.min.ratio}{See \code{glmnet::glmnet()}}

\item{lambda}{User can provide a fixed sequence of lambdas instead
of an automaticaly constructed sequence. If provided, this will override
\code{nlambda} and \code{lambda.min.ratio}}

\item{L}{Square matrix that scales the bias parameter gamma. If \code{NULL}, then
the \code{make_L()} function will be called to set \code{L} as the cholesky
decomposition of \code{crossprod(xtar)}}

\item{standardize}{If TRUE, then \code{xext} will be standardized but only in the
part of the augmented data matrix corresponding to gamma.}

\item{thresh}{See glmnet::glmnet()}

\item{maxit}{See glmnet::glmnet()}
}
\value{
A named list
}
\description{
Use cross-validation to select appropriate value of the tuning parameter in data-enriched generalized linear regression
}
