% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deglr.R
\name{deglr}
\alias{deglr}
\title{Fit data-enriched generalized linear regression model}
\usage{
deglr(
  xtar,
  ytar,
  xext,
  yext,
  family,
  alpha = 0,
  nlambda = 20,
  lambda.min.ratio = ifelse(nrow(xtar) < d/2, 0.001, 1e-06),
  lambda = NULL,
  return_what = NULL,
  L = NULL,
  standardize = TRUE,
  thresh = 1e-14,
  maxit = 1e+05
)
}
\arguments{
\item{xtar}{Predictor matrix from the target population}

\item{ytar}{Outcome vector from the target population}

\item{xext}{Predictor matrix from the external population}

\item{yext}{Outcome vector from the external population}

\item{family}{One of 'gaussian', 'binomial', or 'cox'}

\item{alpha}{See \code{glmnet::cv.glmnet()}}

\item{nlambda}{See glmnet::glmnet() but note that in contrast to glmnet() the
sequence of lambdas will always be exactly this long}

\item{lambda.min.ratio}{See glmnet::glmnet()}

\item{lambda}{User can provide a fixed sequence of lambdas instead
of an automatically constructed sequence. If provided, this will override
\code{nlambda} and \code{lambda.min.ratio}}

\item{return_what}{This argument allows the user to separately indicate that
they only want results returned corresponding to specific values of lambda
or that they only want to calculate an appropriate sequence of lambdas.
This can take on one of three values: \code{NULL} (its default), or a numeric
vector of non-negative numbers, or anything else. If \code{NULL}, then the
function will return the entire solution path. If a numeric vector, then
the function will return the solution path only for these values of lambda.
If anything else, then the function will return a sequence of lambdas and
nothing else.}

\item{L}{Square matrix that scales the bias parameter gamma. If \code{NULL}, then
the \code{make_L()} function will be called to set \code{L} as the cholesky
decomposition of \code{crossprod(xtar)}}

\item{standardize}{If TRUE, then xext will be standardized but only in the
part of the augmented data matrix corresponding to gamma.}

\item{thresh}{See glmnet::glmnet()}

\item{maxit}{See glmnet::glmnet()}
}
\value{
A named list
}
\description{
Fit data-enriched generalized linear regression model
}
