% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1plex.R
\name{log1plex}
\alias{log1plex}
\title{Safe function for accurately calculating log(1 + exp(x)) even for very large and small x's}
\usage{
log1plex(x)
}
\arguments{
\item{x}{Vector of one or more numbers}
}
\value{
A vector of equal length
}
\description{
Safe function for accurately calculating log(1 + exp(x)) even for very large and small x's
}
