% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_L.R
\name{make_L}
\alias{make_L}
\title{Internal helper function for creating default L matrix if not specified}
\usage{
make_L(xtar, family)
}
\arguments{
\item{xtar}{Predictor matrix from the target population}

\item{family}{One of 'gaussian', 'binomial', or 'cox'}
}
\value{
A square matrix with dimension equal to ncol(xtar)
}
\description{
Internal helper function for creating default L matrix if not specified
}
