% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deglr.R
\name{deglr}
\alias{deglr}
\title{Fit data-enriched generalized linear regression model}
\usage{
deglr(
  xtar,
  ytar,
  xext,
  yext,
  family,
  nlambda = 20,
  lambda.min.ratio = 1e-08,
  fixed_lambda_seq = NULL,
  return_only_at = NULL,
  L = NULL,
  standardize = TRUE,
  thresh = 1e-07,
  maxit = 1e+05
)
}
\arguments{
\item{xtar}{Predictor matrix from the target population}

\item{ytar}{Outcome vector from the target population}

\item{xext}{Predictor matrix from the external population}

\item{yext}{Outcome vector from the external population}

\item{family}{One of 'gaussian', 'binomial', or 'cox'}

\item{nlambda}{See glmnet::glmnet() but note that in contrast to glmnet() the
sequence of lambdas will always be exactly this long}

\item{lambda.min.ratio}{See glmnet::glmnet()}

\item{fixed_lambda_seq}{User can provide a fixed sequence of lambdas instead
of an automaticaly constructed sequence. If provided, this will override
nlambda and lambda.min.ratio}

\item{return_only_at}{A vector of lambdas that may or may not have overlap
with fixed_lambda_seq. This argument allows the user to separately indicate
that they only want results returned corresponding to specific values of
lambda.}

\item{L}{Square matrix that scales the bias parameter gamma. The default
value is to use inverse of the cholesky decomposition of crossprod(xtar)}

\item{standardize}{If TRUE, then xext will be standardized but only in the
part of the augmented data matrix corresponding to gamma.}

\item{thresh}{See glmnet::glmnet()}

\item{maxit}{See glmnet::glmnet()}
}
\description{
Fit data-enriched generalized linear regression model
}
